refactor lineage example
